requirements.txt
asgiref==3.3.1
Django==3.1.3
django-rest-auth==0.9.5
djangorestframework==3.12.2
flake8==3.8.4
importlib-metadata==3.1.0
mccabe==0.6.1
pycodestyle==2.6.0
pyflakes==2.2.0
pytz==2020.4
six==1.15.0
sqlparse==0.4.1
zipp==3.4.0
